Sprint 04 back delete slider certifs

+++ src/Controller/AdminController.php
@@ -4,6 +4,7 @@ namespace Beltoise\Controller;
 
 use Beltoise\Model\RenovationManager;
 use Beltoise\Model\RealisationManager;
+use Beltoise\Model\SlideCertificationManager;
 
 class AdminController extends Controller
 {
@@ -16,10 +17,16 @@ class AdminController extends Controller
         $platreries = $realEcoPlatrerieManager->findAllPlatrerie();
         $realEcos = $realEcoPlatrerieManager->findAllRealEco();
 
+        $slideCertificationManager = new SlideCertificationManager();
+        $logos = $slideCertificationManager->findAllLogos();
+        $slides = $slideCertificationManager->findAllSlides();
+
         return $this->twig->render('Admin/admin.html.twig', [
             'renovations' => $renovations,
             'platreries' => $platreries,
             'realEcos' => $realEcos,
+            'logos' => $logos,
+            'slides' => $slides,
         ]);
     }
 
@@ -28,9 +35,12 @@ class AdminController extends Controller
         if (!empty($_POST['id'])) {
             $realisationManager = new RealisationManager();
             $realisation = $realisationManager->find($_POST['id']);
-            $realisationManager->delete($realisation);
-            unlink('assets/uploads/' . $realisation->getImage());
-            header('Location: index.php?route=admin');
+
+            if (file_exists('assets/uploads/' . $realisation->getImage())) {
+                $realisationManager->delete($realisation);
+                unlink('assets/uploads/' . $realisation->getImage());
+                header('Location: index.php?route=admin');
+            }
         }
     }
 
@@ -39,10 +49,42 @@ class AdminController extends Controller
         if (!empty($_POST['id'])) {
             $renovationManager = new RenovationManager();
             $renovation = $renovationManager->find($_POST['id']);
-            $renovationManager->delete($renovation);
-            unlink('assets/uploads/' . $renovation->getImageBefore());
-            unlink('assets/uploads/' . $renovation->getImageAfter());
-            header('Location: index.php?route=admin#AnchorRenovation');
+
+            if (file_exists('assets/uploads/' . $renovation->getImageBefore())
+                and 'assets/uploads/' . $renovation->getImageAfter()) {
+                $renovationManager->delete($renovation);
+                unlink('assets/uploads/' . $renovation->getImageBefore());
+                unlink('assets/uploads/' . $renovation->getImageAfter());
+                header('Location: index.php?route=admin#AnchorRenovation');
+            }
+        }
+    }
+
+    public function deleteSlideAction()
+    {
+        if (!empty($_POST['id'])) {
+            $slideCertificationManager = new SlideCertificationManager();
+            $slide = $slideCertificationManager->find($_POST['id']);
+
+            if (file_exists('assets/uploads/' . $slide->getUri())) {
+                $slideCertificationManager->delete($slide);
+                unlink('assets/uploads/' . $slide->getUri());
+                header('Location: index.php?route=admin#anchorMaconnerie');
+            }
+        }
+    }
+
+    public function deleteCertificationAction()
+    {
+        if (!empty($_POST['id'])) {
+            $slideCertificationManager = new SlideCertificationManager();
+            $certification = $slideCertificationManager->find($_POST['id']);
+
+            if (file_exists('assets/uploads/' . $certification->getUri())) {
+                $slideCertificationManager->delete($certification);
+                unlink('assets/uploads/' . $certification->getUri());
+                header('Location: index.php?route=admin#anchorCertifications');
+            }
         }
     }
 }
